+++ apps/docs/src/index.data.ts
@@ -69,13 +69,19 @@ export default defineLoader({
       } satisfies Partial<ComponentCardProps>;
     };
 
-    const basicComponentsDueDate = new Date(2024, 4);
+    const basicComponentsDueDate = "Q2/2024";
 
     const components: HomePageData["components"] = [
       {
         name: "Button",
         dueDate: basicComponentsDueDate,
         ...getImplementedStatus("OnyxButton"),
+      },
+      {
+        name: "Icon button",
+        dueDate: basicComponentsDueDate,
+        ...getImplementedStatus("OnyxIconButton"),
+        status: "implemented",
       },
       {
         name: "Radio button group",
@@ -92,20 +98,16 @@ export default defineLoader({
         dueDate: basicComponentsDueDate,
         ...getImplementedStatus("OnyxHeadline"),
       },
-      {
-        name: "Footer",
-        dueDate: basicComponentsDueDate,
-        ...getImplementedStatus("OnyxFooter"),
-      },
+
       {
         name: "Header",
         dueDate: basicComponentsDueDate,
         ...getImplementedStatus("OnyxHeader"),
       },
       {
-        name: "Combobox",
+        name: "Select",
         dueDate: basicComponentsDueDate,
-        ...getImplementedStatus("OnyxCombobox"),
+        ...getImplementedStatus("OnyxSelect"),
       },
       {
         name: "Textarea",
@@ -127,6 +129,40 @@ export default defineLoader({
         dueDate: basicComponentsDueDate,
         ...getImplementedStatus("OnyxCheckboxGroup"),
       },
+      {
+        name: "Icons",
+        dueDate: basicComponentsDueDate,
+        status: "implemented",
+        href: "/resources/icons",
+      },
+      {
+        name: "Loading indicator",
+        dueDate: basicComponentsDueDate,
+        ...getImplementedStatus("OnyxLoadingIndicator"),
+        status: "implemented",
+      },
+      {
+        name: "Tooltip",
+        dueDate: basicComponentsDueDate,
+        ...getImplementedStatus("OnyxTooltip"),
+      },
+      {
+        name: "Tag",
+        dueDate: basicComponentsDueDate,
+        ...getImplementedStatus("OnyxTag"),
+      },
+      {
+        name: "Badge",
+        dueDate: basicComponentsDueDate,
+        ...getImplementedStatus("OnyxBadge"),
+      },
+      {
+        name: "Empty",
+        dueDate: basicComponentsDueDate,
+        ...getImplementedStatus("OnyxEmpty"),
+        status: "implemented",
+      },
+      { name: "Footer", ...getImplementedStatus("OnyxFooter") },
       { name: "Advanced Table", ...getImplementedStatus("OnyxTable"), status: "planned" },
       { name: "Filter", ...getImplementedStatus("OnyxFilter") },
       { name: "Notification", ...getImplementedStatus("OnyxNotification") },