+++ apps/playground/src/components/TheHeader.vue
@@ -1,0 +1,122 @@
+<script setup lang="ts">
+import moon from "@sit-onyx/icons/moon.svg?raw";
+import shareAndroid from "@sit-onyx/icons/share-android.svg?raw";
+import sunny from "@sit-onyx/icons/sunny.svg?raw";
+import sync from "@sit-onyx/icons/sync.svg?raw";
+import githubLogo from "../assets/github-logo.svg?raw";
+import HeaderIconButton from "./HeaderIconButton.vue";
+import VersionSelect from "./VersionSelect.vue";
+
+const emit = defineEmits<{
+  reloadPage: [];
+}>();
+
+const isDark = defineModel<boolean>("dark");
+const vueVersion = defineModel<string | null>("vueVersion");
+const typescriptVersion = defineModel<string>("typescriptVersion");
+const onyxVersion = defineModel<string>("onyxVersion");
+
+const copyLink = async () => {
+  await navigator.clipboard.writeText(location.href);
+  alert("URL has been copied to the clipboard.");
+};
+</script>
+
+<template>
+  <header class="header">
+    <a class="header__brand" href="https://onyx.schwarz" target="_blank">
+      <img alt="logo" src="/logo.svg" class="header__logo" />
+      <h1 class="onyx-headline onyx-headline--h2">Playground</h1>
+    </a>
+
+    <div class="header__actions">
+      <VersionSelect
+        v-model="onyxVersion"
+        class="header__version"
+        pkg="sit-onyx"
+        label="onyx version"
+        include-pre-releases
+      />
+
+      <VersionSelect v-model="vueVersion" class="header__version" pkg="vue" label="Vue version" />
+
+      <VersionSelect
+        v-model="typescriptVersion"
+        class="header__version"
+        pkg="typescript"
+        label="TypeScript version"
+      />
+
+      <HeaderIconButton
+        label="Toggle dark mode"
+        :icon="isDark ? moon : sunny"
+        @click="isDark = !isDark"
+      />
+
+      <HeaderIconButton label="Copy URL" :icon="shareAndroid" @click="copyLink" />
+
+      <HeaderIconButton label="Reload page" :icon="sync" @click="emit('reloadPage')" />
+
+      <a
+        href="https://github.com/SchwarzIt/onyx/tree/main/apps/playground"
+        target="_blank"
+        aria-label="View on GitHub"
+      >
+        <HeaderIconButton label="View on GitHub" :icon="githubLogo" variation="secondary" />
+      </a>
+    </div>
+  </header>
+</template>
+
+<style lang="scss" scoped>
+@use "sit-onyx/breakpoints.scss" as onyx;
+
+.header {
+  box-sizing: border-box;
+  padding: var(--onyx-spacing-2xs) var(--onyx-spacing-md);
+  width: 100%;
+  border-bottom: var(--onyx-1px-in-rem) solid var(--onyx-color-base-neutral-200);
+
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: var(--onyx-spacing-sm);
+
+  &__brand {
+    display: flex;
+    align-items: center;
+    gap: var(--onyx-spacing-xs);
+    text-decoration: none;
+  }
+
+  &__logo {
+    height: 1.25rem; // same as name/headline font-size
+  }
+
+  &__actions {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: flex-end;
+    justify-content: flex-end;
+    gap: var(--onyx-spacing-sm) var(--onyx-spacing-md);
+    width: 100%;
+  }
+
+  &__version {
+    width: 10rem;
+  }
+
+  @include onyx.breakpoint(max, md) {
+    flex-direction: column;
+    align-items: flex-start;
+
+    .header__actions {
+      justify-content: flex-start;
+    }
+
+    .header__version {
+      width: 8rem;
+    }
+  }
+}
+</style>