+++ apps/playground/vite.config.ts
@@ -1,0 +1,11 @@
+import vue from "@vitejs/plugin-vue";
+import { defineConfig } from "vite";
+
+// https://vitejs.dev/config
+export default defineConfig({
+  plugins: [vue()],
+  optimizeDeps: {
+    // needed in order for the web workers of "@vue/repl" to work
+    exclude: ["@vue/repl"],
+  },
+});