+++ packages/sit-onyx/src/components/OnyxButton/OnyxButton.ct.tsx
@@ -1,50 +1,69 @@
+import { DENSITIES } from "../../composables/density";
 import { expect, test } from "../../playwright-axe";
-import { executeScreenshotsForAllStates, mockPlaywrightIcon } from "../../utils/playwright";
+import {
+  executeMatrixScreenshotTest,
+  type MatrixScreenshotTestOptions,
+} from "../../playwright/screenshots";
+import { mockPlaywrightIcon } from "../../utils/playwright";
 import OnyxButton from "./OnyxButton.vue";
+import { BUTTON_MODES, BUTTON_VARIATIONS } from "./types";
 
-test("should render", async ({ mount, makeAxeBuilder }) => {
-  // ARRANGE
-  const component = await mount(<OnyxButton label="Button" variation="secondary" />);
+test.describe("Screenshot tests", () => {
+  const screenshotOptions = {
+    rows: ["default", "hover", "active", "focus-visible"] as const,
+    beforeScreenshot: async (component, page, column, row) => {
+      if (row === "hover") await component.hover();
+      if (row === "focus-visible") await page.keyboard.press("Tab");
+      if (row === "active") await page.mouse.down();
+    },
+  } satisfies Partial<MatrixScreenshotTestOptions>;
 
-  // ASSERT
-  await expect(component).toContainText("Button");
+  for (const mode of BUTTON_MODES) {
+    executeMatrixScreenshotTest({
+      ...screenshotOptions,
+      columns: BUTTON_VARIATIONS,
+      name: `Button (${mode})`,
+      component: (column) => <OnyxButton label="Button" mode={mode} variation={column} />,
+    });
 
-  // ACT
-  const accessibilityScanResults = await makeAxeBuilder().analyze();
+    executeMatrixScreenshotTest({
+      ...screenshotOptions,
+      columns: BUTTON_VARIATIONS,
+      name: `Button (${mode}, disabled)`,
+      component: (column) => <OnyxButton label="Button" mode={mode} variation={column} disabled />,
+    });
 
-  // ASSERT
-  expect(accessibilityScanResults.violations).toEqual([]);
-});
+    executeMatrixScreenshotTest({
+      ...screenshotOptions,
+      columns: BUTTON_VARIATIONS,
+      name: `Button (${mode}, with icon)`,
+      component: (column) => (
+        <OnyxButton label="Button" mode={mode} variation={column} icon={mockPlaywrightIcon} />
+      ),
+    });
 
-test("should display correctly when disabled", async ({ mount, makeAxeBuilder }) => {
-  // ARRANGE
-  const component = await mount(<OnyxButton label="Button" variation="secondary" disabled />);
+    executeMatrixScreenshotTest({
+      ...screenshotOptions,
+      columns: BUTTON_VARIATIONS,
+      name: `Button (${mode}, loading)`,
+      component: (column) => <OnyxButton label="Button" mode={mode} variation={column} loading />,
+    });
+  }
 
-  // ASSERT
-  await expect(component).toContainText("Button");
-  await expect(component).toBeDisabled();
-
-  // ACT
-  const accessibilityScanResults = await makeAxeBuilder().analyze();
-
-  // ASSERT
-  expect(accessibilityScanResults.violations).toEqual([]);
-});
-
-test("should render button with icon", async ({ mount, makeAxeBuilder }) => {
-  // ARRANGE
-  const component = await mount(
-    <OnyxButton label="Button" variation="secondary" icon={mockPlaywrightIcon} />,
-  );
-
-  // ASSERT
-  await expect(component).toContainText("Button");
-
-  // ACT
-  const accessibilityScanResults = await makeAxeBuilder().analyze();
-
-  // ASSERT
-  expect(accessibilityScanResults.violations).toEqual([]);
+  executeMatrixScreenshotTest({
+    ...screenshotOptions,
+    columns: DENSITIES,
+    rows: [...BUTTON_MODES, "skeleton"],
+    name: `Button (densities)`,
+    component: (column, row) => (
+      <OnyxButton
+        label="Button"
+        density={column}
+        mode={row !== "skeleton" ? row : undefined}
+        skeleton={row === "skeleton"}
+      />
+    ),
+  });
 });
 
 test("should truncate text", async ({ mount }) => {
@@ -63,7 +82,10 @@ test("should truncate text", async ({ mount }) => {
 test("should render button inline aligned with text", async ({ mount }) => {
   // ARRANGE
   const component = await mount(
-    <p class="onyx-text" style="line-height: 60px; width: max-content;">
+    <p
+      class="onyx-text"
+      style="line-height: 60px; width: max-content; font-family: var(--onyx-font-family);"
+    >
       before
       <OnyxButton label="Test label" />
       between
@@ -74,72 +96,4 @@ test("should render button inline aligned with text", async ({ mount }) => {
 
   // ASSERT
   await expect(component).toHaveScreenshot("inline-aligned.png");
-});
-
-test("should render correct inline", async ({ mount }) => {
-  // ARRANGE
-  const component = await mount(<OnyxButton label="Test label" skeleton />);
-
-  // ASSERT
-  await expect(component).toHaveScreenshot("skeleton.png");
-});
-
-const STATES = {
-  state: ["default", "disabled", "icon"],
-  variation: ["primary", "secondary", "danger"],
-  mode: ["default", "outline", "plain"],
-  focusState: ["", "hover", "focus-visible", "active"],
-} as const;
-
-test.describe("state screenshot tests", () => {
-  executeScreenshotsForAllStates(
-    STATES,
-    "button",
-    async ({ variation, state, mode, focusState }, mount, page) => {
-      const component = await mount(
-        <OnyxButton
-          label="label"
-          variation={variation}
-          mode={mode}
-          disabled={state === "disabled"}
-          icon={state === "icon" ? mockPlaywrightIcon : undefined}
-        />,
-      );
-
-      const button = component.getByRole("button");
-
-      if (focusState === "focus-visible") await page.keyboard.press("Tab");
-      if (focusState === "hover") await button.hover();
-      if (focusState === "active") await page.mouse.down();
-
-      return component;
-    },
-  );
-});
-
-const DENSITYSTATES = {
-  density: ["compact", "default", "cozy"],
-  focusState: ["focus-visible"],
-} as const;
-
-test.describe("state density screenshot tests", () => {
-  executeScreenshotsForAllStates(
-    DENSITYSTATES,
-    "button",
-    async ({ density, focusState }, mount, page) => {
-      const component = await mount(
-        <OnyxButton
-          label="label"
-          variation="primary"
-          mode="default"
-          density={density}
-          icon={mockPlaywrightIcon}
-        />,
-      );
-
-      if (focusState === "focus-visible") await page.keyboard.press("Tab");
-
-      return component;
-    },
-  );
 });