+++ packages/sit-onyx/src/components/OnyxListbox/types.ts
@@ -19,6 +19,15 @@ export type OnyxListboxProps<TValue extends SelectionOptionValue = SelectionOpti
    * Current value / selected options.
    */
   modelValue?: TValue;
+  /**
+   * Whether to show a loading indicator.
+   */
+  loading?: boolean;
+  /**
+   * Lazy loading options. Can be used to load more options on scroll.
+   * If you want to use a button instead, use the `optionsEnd` slot.
+   */
+  lazyLoading?: ListboxLazyLoading;
 };
 
 export type ListboxOption<T extends SelectionOptionValue = SelectionOptionValue> = {
@@ -34,4 +43,21 @@ export type ListboxOption<T extends SelectionOptionValue = SelectionOptionValue>
    * Whether the option is disabled.
    */
   disabled?: boolean;
+};
+
+export type ListboxLazyLoading = {
+  /**
+   * Whether lazy loading should be enabled.
+   * Can be disabled e.g. if all options are loaded.
+   */
+  enabled?: boolean;
+  /**
+   * Whether the lazy loading is currently loading more options.
+   */
+  loading?: boolean;
+  /**
+   * Scroll offset (in pixel). Must be >= 0.
+   * Can be used to trigger the `lazyLoad` event earlier (e.g. if scrolled to second last option).
+   */
+  scrollOffset?: number;
 };