+++ packages/sit-onyx/src/composables/density.ts
@@ -1,14 +1,14 @@
 import { computed } from "vue";
 
-const DENSITY = ["cozy", "default", "compact"] as const;
-export type DensityType = (typeof DENSITY)[number];
+export const DENSITIES = ["compact", "default", "cozy"] as const;
+export type Density = (typeof DENSITIES)[number];
 
 export type DensityProp = {
   /**
    * Density defines the amount of vertical white space a component has and the height of the main (interactive) element of a component.
    * @default undefined By default the parents setting is used, if none is defined on any parent `default` is the default.
    */
-  density?: DensityType;
+  density?: Density;
 };
 
 export const useDensity = (props: DensityProp) => ({